write proxy in .json file

+++ test/test.py
@@ -1,14 +1,17 @@
 #!/usr/bin/env python3
 
 import os
+import sys
 
-f = open('Makefile', 'w')
-seq = '''a.pdf:
-\tlatexmk -xelatex -pvc- -shell-escape a.tex
-'''
-f.write(seq)
-f.close()
-os.system('python ../AutoTrans.py a.txt')
-os.system('make')
-os.system('cp temp/a.pdf .')
-os.system('rm -rf temp')
+opt = sys.argv[1]
+
+if opt == '-t' or opt == '--translate':
+    os.system('python ../AutoTrans.py a.txt')
+    os.system('make')
+    os.system('cp temp/a.pdf .')
+    os.system('rm -rf temp')
+elif opt == '-p' or opt == '--proxy':
+    sys.path.append("./..")
+    sys.path.append(".")
+    from Proxy import proxy
+    proxy.proxy_get()
